Add more test coverage for logic that uses delegateMinHeightActive value - Closes #4499

+++ framework/src/modules/chain/components/storage/migrations/sql/20190522152200_add_round_delegates_table.sql
@@ -20,6 +20,6 @@
 
 /* Table */
 CREATE TABLE IF NOT EXISTS "round_delegates" (
-  "round" BIGINT NOT NULL PRIMARY KEY,
+  "round" INT NOT NULL PRIMARY KEY,
   "delegatePublicKeys" JSON NOT NULL
 );
